Add files via upload
I have created a list of tests to run to check the basic functionality of the lifts and special requirements. I have described the actions and not just the elements I want to test. I analyzed the functionality of the ordinary elevator, I described the tests in a clear and structured way.